refactor: split ai_handler.rs into focused modules (closes #473)
Extract build_context_prompt into context_builder.rs and approval
helper functions (tool_requires_approval, normalize_relative_path,
build_approval_rule) into approval_rules.rs.

ai_handler.rs: 1279 → 1163 lines
