+++ server/src/handlers/ai_handler.rs
@@ -19,90 +19,13 @@ use tokio::time::{timeout, Duration};
 use super::common::{
     build_approval_request_payload, build_streaming_payload, error_response, now_millis,
     tool_log_from_call, with_system_prompts, AIMode, AgentEventPayload, AgentEventStatus, AppState,
-    ApprovalDecisionPayload, ApprovalOption, ApprovalRule, ArtifactDetailsPayload, ArtifactKind,
+    ApprovalDecisionPayload, ApprovalOption, ArtifactDetailsPayload, ArtifactKind,
     PendingEditPayload, PlanStepKind, PlanStepPayload, PlanStepStatus, ToolLogStatus,
     ToolPreviewPayload, WSResponse,
 };
+use super::approval_rules::{build_approval_rule, normalize_relative_path, tool_requires_approval};
+use super::context_builder::build_context_prompt;
 use super::tool_handler::execute_ai_tool_call;
-
-pub(super) async fn build_context_prompt(
-    runtime: &Arc<ProjectRuntime>,
-    context: reprod_core::acp::types::AcpContextRequest,
-) -> Result<String, anyhow::Error> {
-    let mut parts = Vec::new();
-
-    // 1. Console Context
-    if let Some(limit) = context.console_history_limit {
-        if limit > 0 {
-            let runs = runtime
-                .execution_repo
-                .latest_runs(limit)
-                .await
-                .unwrap_or_default();
-
-            if !runs.is_empty() {
-                let mut console_text =
-                    String::from("Recent console output (newest first, truncated):\n");
-                for run in runs {
-                    let status = format!("{:?}", run.status).to_lowercase();
-                    let duration = run.duration_ms.unwrap_or(0);
-                    console_text.push_str(&format!(
-                        "- [{}] {} in {}ms\n",
-                        run.created_at_ms, status, duration
-                    ));
-                    if !run.result.output.is_empty() {
-                        let trimmed = run
-                            .result
-                            .output
-                            .lines()
-                            .take(20)
-                            .collect::<Vec<_>>()
-                            .join("\n");
-                        let truncated = if trimmed.len() > 800 {
-                            &trimmed[..800]
-                        } else {
-                            &trimmed
-                        };
-                        console_text.push_str(&format!("stdout: {}\n", truncated));
-                    }
-                    if let Some(err) = &run.result.error {
-                        let trimmed = err.lines().take(20).collect::<Vec<_>>().join("\n");
-                        let truncated = if trimmed.len() > 800 {
-                            &trimmed[..800]
-                        } else {
-                            &trimmed
-                        };
-                        console_text.push_str(&format!("stderr: {}\n", truncated));
-                    }
-                    console_text.push('\n');
-                }
-                parts.push(console_text);
-            }
-        }
-    }
-
-    // 2. File Context
-    if let Some(path) = context.active_buffer_path {
-        if !path.is_empty() {
-            match runtime.edit_service.read_text_file(&path).await {
-                Ok(result) => {
-                    parts.push(format!(
-                        "Current file ({}):\n\n```r\n{}\n```\n",
-                        path, result.text
-                    ));
-                }
-                Err(e) => {
-                    tracing::warn!("Failed to read context file {}: {}", path, e);
-                }
-            }
-        }
-    }
-
-    // 3. User Input
-    parts.push(context.user_input);
-
-    Ok(parts.join("\n"))
-}
 
 type ResponseSender = Option<UnboundedSender<WSResponse>>;
 
@@ -150,10 +73,6 @@ impl EventStream {
             },
         );
     }
-}
-
-fn tool_requires_approval(name: &str) -> bool {
-    matches!(name, "apply_pending_edit")
 }
 
 fn push_cancel_response(
@@ -326,41 +245,6 @@ fn mark_plan_finished(
         return true;
     }
     false
-}
-
-fn normalize_relative_path(path: &str) -> Option<String> {
-    use std::path::Component;
-    let mut parts = Vec::new();
-    let path = std::path::Path::new(path);
-    for component in path.components() {
-        match component {
-            Component::Normal(part) => parts.push(part.to_string_lossy().to_string()),
-            Component::CurDir => {}
-            _ => return None,
-        }
-    }
-    if parts.is_empty() {
-        return None;
-    }
-    Some(parts.join("/"))
-}
-
-fn approval_prefix_from_path(path: &str) -> Option<String> {
-    let normalized = normalize_relative_path(path)?;
-    if let Some((parent, _)) = normalized.rsplit_once('/') {
-        if !parent.is_empty() {
-            return Some(parent.to_string());
-        }
-    }
-    Some(normalized)
-}
-
-fn build_approval_rule(tool: &str, path: Option<&str>) -> Option<ApprovalRule> {
-    let path_prefix = path.and_then(approval_prefix_from_path);
-    Some(ApprovalRule {
-        tool: tool.to_string(),
-        path_prefix,
-    })
 }
 
 fn build_diff(path: &str, old_text: &str, new_text: &str) -> String {
